Oops, didnt include all of teh changes.
I added the generated Content Provider and incorporated the DataDroid library for use in easily passing JSON formatted data between the phones and the server.

Issues: the logic is in place to query the server for a bundle of sample Vitals data. An array of Vitals is successfully transmitted form the server (delta) to the phone, but during the process of adding the points to the chart series it hangs, so I will have to fix this by just querying the database for the data instead of trying to use the query result, as the results are dumped into the database anyway during the parsing of the JSON result.

Please see the demo projects associated with these two projects for examples on how to interface with the content provider, or you can google it as the content provider pattern is half built into the android api so it's implemented and used quite a lot.

https://github.com/foxykeep/DataDroid
https://github.com/foxykeep/ContentProviderCodeGenerator