I added the functionality to allow for passing parameters to the broker's vitals query servlet. This includes the requested patient id, an initial vid, and a limit of entries to return.
The limit is currently set to 10 by default.

For testing purposes the patient id is initialized to 1 but should be initialized to a null or a negative number before a patient is selected either implicitly of explicitly from the banner view.

The content provider insert function was modified to overwrite sql constraint conflicts. This was added to ensure that new vitals entries are overwritten when a duplicate (pid , vid) pair is submitted to be inserted on the local android database. In order to avoid loss of historic data duplicate entries should be properly managed on the broker.

Polling for vitals entries is performed every 1/4 second. entries corresponding to vids older than the latest plotted entry are discarded. And only entries specific to the current patient id are plotted.

In the current state of the app the graph is quickly overloaded with data points and causes the app to freeze. Fixing this will require the use of a sliding window to cap the maximum amount of visualized data points to around 300 or 30 ecg waveform phases.