The issue encountered yesterday regarding a null reference to the object filed of the message passed from the UDPClient to the Patient fragment message handler was being caused by the received message being used by the system for another purpose. This issue was solved by creating a separate and final reference to the message object before the plotting routine is started.
Additionally, the messaging communication scheme has been improved by transmitting a java array of parcelable Vital objects instead of sending them in series. This modification slightly reduces the transmission payload overhead by relying on a simpler container.

Moreover, the integer value encoded in the head of the data stream, representing the number of vitals in the stream is no longer needed as the size is encoded in the array object, however it has been left in for debugging purposes.

Relying on the UDP for sending batches of ecg data has resulted, as expected in the occasional loss of sections of the waveform. So far, at most only a single bunch of data points are lost, resulting in about a quarter second gap in the waveform. Due to the current size of data bunch (50 samples) this gap of a quarter second is also the minimum gap which can be caused by a dropped message.

In an effort to reduce the amount of work being offloaded to the garbage collector, a dedicated thread has been added for the purpose of plotting the ecg data points, rather than creating a new runnable each time a message is received.