Banner now updates patients using MQTT messages.
MQTT service manager placed in Main Activity so it can pass messages to the Banner and/or PatientLeft as needed and the JSON is parsed just once.

Patient now has a src field used in place of a patient ID, which the current MQTT record topic does not have.

Added Service to Manifest.xml