Updated MQTT library in attempt to fix connection issue (may not be needed).
switched banner to support fragment to see if startup of Scenario activity lag was from two different fragment types. Seems to be from the service starting (not from the call to start service though).

removed custom time to wait in mqtt client (again to try to fix connection issue)

launch old UI with extra button (for testing).

sorry for all the small changes in one commit.